refactor: boardService 추상화
